Bring back WASMTranslationHelper by loading it in an iframe in the main content
Trying to use Worker in contentScript did not give me the right origin, but loading an iframe does? Anyway, this way we forward all requests from all frames to the background script, and then background script will forward them to a single iframe that hosts a WASMTranslationHelper. And it works! There are some opportunities to use makeServer and makeClient in other places as well I think, because this is about the 5th time I've implemented that functionality…